+++ types/v56/nsxt_types.go
@@ -182,3 +182,7 @@ type Connection struct {
 	RouterRef      OpenApiReference `json:"routerRef"`
 	ConnectionType string           `json:"connectionType,omitempty"`
 }
+
+// NsxtLogicalSwitch is a type alias with better name for holding NSX-T Logical Switches which can be used to back NSX-T
+// imported Org Vdc network
+type NsxtLogicalSwitch = OpenApiReference